Refactoring

+++ src/app/config.h
@@ -65,21 +65,18 @@ struct __attribute ((packed)) SysConfig {
 
 struct __attribute ((packed)) NightModeConfig {
     bool enabled = false;
-
     uint16_t brightness = 10;
-
     uint32_t start_time = 0;
-    uint32_t end_time = (uint32_t) 10 * 60 * 60;
-    uint16_t switch_interval = (uint32_t) 15 * 60;
+    uint32_t end_time = 10 * 60 * 60; // 22:00
+    uint16_t switch_interval = 15 * 60; // 15 minutes
 };
 
 struct __attribute ((packed)) Config {
     bool power = true;
-
     uint16_t brightness = 2048;
 
-    uint32_t color = ~0u;
-    uint32_t calibration = ~0u;
+    uint32_t color = ~0u; // All colors
+    uint32_t calibration = ~0u; // No calibration
 
     NightModeConfig night_mode {};
